Rewrite action from Rust to Go, to minimize third-party dependencies
The original motivation for writing this GitHub action myself was to
minimize third-party dependencies. But when writing Rust, one has to
depend on third-party libraries, since the standard library is
relatively light-weight. Go, on the other hand, has a batteries-included
standard library, letting us build the action without any dependencies
besides my own utility libraries for Go.

In addition, the greater stability of Go compared to Rust is nice, since
I want the burden of maintaining this action to be as low as possible.